# Modules/CustomerPortalApi/OpenApi/openapi.yaml
openapi: 3.1.0
info:
  title: New World Cargo Customer Portal API
  version: 1.0.0
  description: Versioned customer-owned API for the separate React customer island.
servers:
  - url: https://api.example.com/api/v1
    description: Configure per environment.
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: {type: string, maxLength: 255}
                password: {type: string, minLength: 1}
      responses:
        '200':
          description: Customer session created.
          headers:
            Set-Cookie: {schema: {type: string}}
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiSuccessAuthUser'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, email, phone, password]
              properties:
                firstName: {type: string, minLength: 2, maxLength: 80}
                lastName: {type: string, maxLength: 80}
                email: {type: string, format: email}
                phone: {type: string, maxLength: 30}
                password: {type: string, minLength: 8}
      responses:
        '201':
          description: Customer account created.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiSuccessAuthUser'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
  /auth/verify:
    post:
      tags: [Auth]
      operationId: verify
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: {type: string, pattern: '^[0-9]{6}$'}
      responses:
        '200':
          description: Contact verified.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiSuccessNull'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
  /session:
    get:
      tags: [Auth]
      security: [{cookieAuth: []}]
      operationId: session
      responses:
        '200':
          description: Current authenticated customer.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiSuccessAuthUser'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
...
  /returns:
    get:
      tags: [Returns]
      security: [{cookieAuth: []}]
      operationId: returns
      responses:
        '200':
          description: Customer return requests.
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiSuccess'}
                  - type: object
                    properties:
                      data: {type: array, items: {$ref: '#/components/schemas/ReturnRequestDto'}}
    post:
      tags: [Returns]
      security: [{cookieAuth: [], csrfHeader: []}]
      operationId: createReturn
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReturnRequestInput'}
      responses:
        '201':
          description: Return request created.
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/ApiSuccess'}
                  - type: object
                    properties:
                      data: {$ref: '#/components/schemas/ReturnRequestDto'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: laravel_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  headers:
    RequestId:
      description: Correlation identifier echoed by the API.
      schema: {type: string}
  schemas:
    ApiSuccess:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        meta: {type: object, additionalProperties: true}
        requestId: {type: string}
...
    ApiSuccessAuthUser:
      allOf:
        - {$ref: '#/components/schemas/ApiSuccess'}
        - type: object
          properties:
            data: {$ref: '#/components/schemas/AuthUser'}
    ApiSuccessNull:
      allOf:
        - {$ref: '#/components/schemas/ApiSuccess'}
        - type: object
          properties: {data: {type: 'null'}}
...
    ReturnRequestInput:
      type: object
      required: [shipmentId, reason, handover]
      properties:
        shipmentId: {type: integer}
        reason: {type: string}
        handover: {type: string, enum: [pickup, drop_off]}
    ReturnRequestDto:
      allOf:
        - {$ref: '#/components/schemas/ReturnRequestInput'}
        - type: object
          required: [id, customerId, trackingNumber, status, revision]
          properties:
            id: {type: string}
            customerId: {type: string}
            trackingNumber: {type: string, nullable: true}
            status: {type: string}
            displayStatus: {type: string}
            createdAt: {type: string, format: date-time, nullable: true}
            revision: {type: integer, minimum: 1}
...
    AuthUser:
      type: object
      required: [id, firstName, lastName, email, phone, provider, verified]
      properties:
        id: {type: string}
        firstName: {type: string}
        lastName: {type: string}
        email: {type: string, format: email}
        phone: {type: string, nullable: true}
        avatar: {type: string, nullable: true}
        provider: {type: string, enum: [password, google]}
        verified: {type: boolean}
...
    Problem:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            fieldErrors: {type: object, additionalProperties: {type: array, items: {type: string}}}
            retryable: {type: boolean}
        requestId: {type: string}
  responses:
...
    Unauthenticated:
      description: No valid session.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: Authenticated account cannot use this portal.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Problem'}
    ValidationFailed:
      description: Validation failed.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Problem'}
...